# contracts/openapi/w2-stage10a-openapi.json
{
  "components": {
    "schemas": {
...
      "FixtureDetailResponse": {
        "properties": {
          "away_team_id": {
            "title": "Away Team Id",
            "type": "string"
          },
          "bookmaker_count": {
            "title": "Bookmaker Count",
            "type": "integer"
          },
          "competition_id": {
            "title": "Competition Id",
            "type": "string"
          },
          "competition_name": {
            "title": "Competition Name",
            "type": "string"
          },
          "data_state": {
            "title": "Data State",
            "type": "string"
          },
          "fixture_id": {
            "title": "Fixture Id",
            "type": "string"
          },
          "forward_decision": {
            "title": "Forward Decision",
            "type": "string"
          },
          "home_team_id": {
            "title": "Home Team Id",
            "type": "string"
          },
          "kickoff_display": {
            "title": "Kickoff Display",
            "type": "string"
          },
          "kickoff_utc": {
            "format": "date-time",
            "title": "Kickoff Utc",
            "type": "string"
          },
          "lifecycle_state": {
            "title": "Lifecycle State",
            "type": "string"
          },
          "market_coverage": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Market Coverage",
            "type": "object"
          },
          "provenance": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Provenance",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "risk_notes": {
            "items": {
              "type": "string"
            },
            "title": "Risk Notes",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          }
        },
        "required": [
          "fixture_id",
          "competition_id",
          "competition_name",
          "kickoff_utc",
          "kickoff_display",
          "status",
          "home_team_id",
          "away_team_id",
          "lifecycle_state",
          "data_state",
          "request_id",
          "venue",
          "bookmaker_count",
          "market_coverage",
          "forward_decision",
          "provenance",
          "risk_notes"
        ],
        "title": "FixtureDetailResponse",
        "type": "object"
      },
      "FixtureListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FixtureSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "meta",
          "items"
        ],
        "title": "FixtureListResponse",
        "type": "object"
      },
      "FixtureSummary": {
        "properties": {
          "away_team_id": {
            "title": "Away Team Id",
            "type": "string"
          },
          "competition_id": {
            "title": "Competition Id",
            "type": "string"
          },
          "competition_name": {
            "title": "Competition Name",
            "type": "string"
          },
          "data_state": {
            "title": "Data State",
            "type": "string"
          },
          "fixture_id": {
            "title": "Fixture Id",
            "type": "string"
          },
          "home_team_id": {
            "title": "Home Team Id",
            "type": "string"
          },
          "kickoff_display": {
            "title": "Kickoff Display",
            "type": "string"
          },
          "kickoff_utc": {
            "format": "date-time",
            "title": "Kickoff Utc",
            "type": "string"
          },
          "lifecycle_state": {
            "title": "Lifecycle State",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "fixture_id",
          "competition_id",
          "competition_name",
          "kickoff_utc",
          "kickoff_display",
          "status",
          "home_team_id",
          "away_team_id",
          "lifecycle_state",
          "data_state"
        ],
        "title": "FixtureSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OddsPoint": {
        "properties": {
          "bookmaker_count": {
            "title": "Bookmaker Count",
            "type": "integer"
          },
          "captured_at": {
            "format": "date-time",
            "title": "Captured At",
            "type": "string"
          },
          "closing": {
            "title": "Closing",
            "type": "boolean"
          },
          "decimal_odds": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decimal Odds"
          },
          "first_seen": {
            "title": "First Seen",
            "type": "boolean"
          },
          "line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "selection": {
            "title": "Selection",
            "type": "string"
          },
          "snapshot_semantics": {
            "title": "Snapshot Semantics",
            "type": "string"
          }
        },
        "required": [
          "captured_at",
          "snapshot_semantics",
          "market",
          "selection",
          "line",
          "decimal_odds",
          "bookmaker_count",
          "first_seen",
          "closing"
        ],
        "title": "OddsPoint",
        "type": "object"
      },
      "OddsTimelineResponse": {
        "properties": {
          "fixture_id": {
            "title": "Fixture Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OddsPoint"
            },
            "title": "Items",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "fixture_id",
          "items"
        ],
        "title": "OddsTimelineResponse",
        "type": "object"
      },
...
      "PageMeta": {
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total"
        ],
        "title": "PageMeta",
        "type": "object"
      },
      "ProbabilityResponse": {
        "properties": {
          "as_of_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Time"
          },
          "calibrated": {
            "default": false,
            "title": "Calibrated",
            "type": "boolean"
          },
          "fixture_id": {
            "title": "Fixture Id",
            "type": "string"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Probabilities",
            "type": "object"
          },
          "probability_type": {
            "title": "Probability Type",
            "type": "string"
          },
          "quality": {
            "title": "Quality",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "fixture_id",
          "probability_type",
          "probabilities",
          "source",
          "quality"
        ],
        "title": "ProbabilityResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "W2 Football Intelligence Engine",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/fixtures": {
      "get": {
        "operationId": "list_fixtures_v1_fixtures_get",
        "parameters": [
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "competition_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Competition Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "team_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Id"
            }
          },
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "UTC",
              "title": "Timezone",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "If-None-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixtureListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Fixtures",
        "tags": [
          "public-read"
        ]
      }
    },
    "/v1/fixtures/{fixture_id}": {
      "get": {
        "operationId": "fixture_detail_v1_fixtures__fixture_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "fixture_id",
            "required": true,
            "schema": {
              "title": "Fixture Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "UTC",
              "title": "Timezone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixtureDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fixture Detail",
        "tags": [
          "public-read"
        ]
      }
    },
    "/v1/fixtures/{fixture_id}/market-probabilities": {
      "get": {
        "operationId": "market_probabilities_v1_fixtures__fixture_id__market_probabilities_get",
        "parameters": [
          {
            "in": "path",
            "name": "fixture_id",
            "required": true,
            "schema": {
              "title": "Fixture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProbabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Probabilities",
        "tags": [
          "public-read"
        ]
      }
    },
    "/v1/fixtures/{fixture_id}/model-probabilities": {
      "get": {
        "operationId": "model_probabilities_v1_fixtures__fixture_id__model_probabilities_get",
        "parameters": [
          {
            "in": "path",
            "name": "fixture_id",
            "required": true,
            "schema": {
              "title": "Fixture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProbabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Model Probabilities",
        "tags": [
          "public-read"
        ]
      }
    },
    "/v1/fixtures/{fixture_id}/odds-timeline": {
      "get": {
        "operationId": "odds_timeline_v1_fixtures__fixture_id__odds_timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "fixture_id",
            "required": true,
            "schema": {
              "title": "Fixture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OddsTimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Odds Timeline",
        "tags": [
          "public-read"
        ]
      }
    },
...
  }
}